test: split coordinator update-cycle tests further
